requirements.txt
Flask
Gunicorn
requests==2.26.0
python-decouple==3.5